requirements.txt
streamlit
chromadb
google-generativeai
plotly
pandas